requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-mail==1.4.1
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
instructor==1.4.3
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
multidict==6.1.0
openai==1.46.0
packaging==24.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
redis==5.1.1
requests==2.32.3
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.13.0
